# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Oficina Backend API",
        "description": "Documentacao base da API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reports/os-summary": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Resumo geral das Ordens de Serviço: contagem por status e receita total",
                "operationId": "b1740ffbb2268759cb14a1b55a1a3483",
                "responses": {
                    "200": {
                        "description": "Resumo de OS"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/revenue": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Receita de OS finalizadas em um período (from/to em YYYY-MM-DD)",
                "operationId": "c245ffd5822d7f98ca6dbfc3ef815497",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-12-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receita no período"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/avg-execution-time": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Tempo médio de execução das OS (do início ao fim da execução)",
                "operationId": "a5dfe0490c6f54432cf43e49050d1f6e",
                "responses": {
                    "200": {
                        "description": "Tempo médio de execução"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/low-stock": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Peças com estoque abaixo ou igual ao estoque mínimo",
                "operationId": "13316ca093e4962fa3cd329e664018c7",
                "responses": {
                    "200": {
                        "description": "Lista de peças com estoque baixo"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token gerado pelo endpoint POST /api/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}